Fixed some bugs

+++ domains/Account/Database/Migrations/2017_12_24_054838_create_user_phones_table.php
@@ -15,7 +15,7 @@ class CreateUserPhonesTable extends Migration
     {
         Schema::create('user_phones', function (Blueprint $table) {
             $table->unsignedInteger('user_id');
-            $table->integer('number');
+            $table->bigInteger('number');
             $table->foreign('user_id')
                 ->references('id')
                 ->on('users')
